+++ pkg/rabbitmqamqp/requester.go
@@ -123,9 +123,11 @@ type RequesterOptions struct {
 	// Optional. If not set, a default timeout of 30 seconds will be used.
 	RequestTimeout time.Duration
 
-	// If true, the requester will set the 'Direct-Reply-To' feature for RabbitMQ.
-	// see: https://www.rabbitmq.com/direct-reply-to.html
-	DirectReplyTo bool
+	// SettleStrategy configures how the reply consumer receives messages.
+	// Use ExplicitSettle for a dedicated reply queue (default).
+	// Use DirectReplyTo to enable RabbitMQ direct-reply-to (no reply queue declared).
+	// See: https://www.rabbitmq.com/docs/direct-reply-to#overview
+	SettleStrategy ConsumerSettleStrategy
 }
 
 type outstandingRequest struct {